add check

+++ app/main.py
@@ -2,13 +2,16 @@ import os
 
 
 def move_file(command: str) -> None:
-    command = command.split()
-    cmd, old_loc, new_loc = command
-    new_directory = os.path.dirname(new_loc)
-    if new_directory:
-        os.makedirs(new_directory, exist_ok=True)
-    with (open(old_loc, "r") as source_file,
-            open(new_loc, "w") as destination_file):
-        destination_file.write(source_file.read())
-    if os.path.exists(old_loc):
-        os.remove(old_loc)
+    if len(parts) == 3 and parts[0] == "mv":
+        command = command.split()
+        cmd, old_loc, new_loc = command
+        new_directory = os.path.dirname(new_loc)
+        if new_directory:
+            os.makedirs(new_directory, exist_ok=True)
+        with (open(old_loc, "r") as source_file,
+                open(new_loc, "w") as destination_file):
+            destination_file.write(source_file.read())
+        if os.path.exists(old_loc):
+            os.remove(old_loc)
+    else:
+        raise ValueError("Invalid command")
